+++ packages/mg-components/src/components/molecules/mg-form/mg-form.tsx
@@ -115,21 +115,38 @@ export class MgForm {
     this.classCollection.delete(this.classAllRequired);
     // If the form is disabled or readonly none of them are required
     // Check if all fields are not editable (readonly or disabled)
-    if (!this.disabled && !this.readonly && !this.mgInputs.every(input => input.disabled || input.readonly)) {
-      // Get required fields
-      const requiredInputs = this.mgInputs.filter(input => input.required && !input.disabled && !input.readonly);
-      // All fields are required
+    const isEditable = input => !(input.disabled || input.readonly);
+    const isMgInputToggle = input => input.nodeName === 'MG-INPUT-TOGGLE';
+    if (!this.disabled && !this.readonly && this.mgInputs.some(input => isEditable(input) && !isMgInputToggle(input))) {
+      // Get editable inputs
+      const editableInputs: HTMLMgInputsElement[] = this.mgInputs.filter(isEditable);
+      // Get required inputs
       // mg-input-toggle can not be required
-      if (requiredInputs.length > 0 && requiredInputs.length === this.mgInputs.filter(input => input.nodeName !== 'MG-INPUT-TOGGLE').length) {
-        this.requiredMessage = requiredInputs.length === 1 ? this.messages.form.allRequiredSingle : this.messages.form.allRequired;
+      const requiredInputs: HTMLMgInputsElement[] = editableInputs.filter(input => !isMgInputToggle(input) && input.required);
+      // All inputs are required
+      if (
+        [requiredInputs.length, editableInputs.length].every(length => length === 1) ||
+        (requiredInputs.length > 1 && requiredInputs.length === editableInputs.filter(input => !isMgInputToggle(input)).length)
+      ) {
+        this.requiredMessage = this.getRequiredMessageBasedOnCount(requiredInputs, 'allRequiredSingle', 'allRequired');
         this.classCollection.add(this.classAllRequired);
       }
       // Some fields are required
       else if (requiredInputs.length > 0) {
-        this.requiredMessage = requiredInputs.length === 1 ? this.messages.form.requiredSingle : this.messages.form.required;
+        this.requiredMessage = this.getRequiredMessageBasedOnCount(requiredInputs, 'requiredSingle', 'required');
       }
     }
   };
+
+  /**
+   * Returns a required message based on the count of required inputs.
+   * @param requiredInputs - An array of required input elements.
+   * @param keySingle - The key for the singular message format.
+   * @param keyPlural - The key for the plural message format.
+   * @returns The appropriate required message based on the count of required inputs.
+   */
+  private getRequiredMessageBasedOnCount = (requiredInputs: HTMLMgInputsElement[], keySingle: string, keyPlural: string) =>
+    this.messages.form[requiredInputs.length === 1 ? keySingle : keyPlural];
 
   /**
    * Check if form is valid
@@ -229,7 +246,7 @@ export class MgForm {
    */
   render(): HTMLElement {
     return (
-      <form class={this.classCollection.join()} id={this.identifier} name={this.name} ref={el => (this.form = el as HTMLFormElement)} onSubmit={this.handleFormSubmit}>
+      <form class={this.classCollection.join()} id={this.identifier} name={this.name} ref={(el: HTMLFormElement) => (this.form = el)} onSubmit={this.handleFormSubmit}>
         {this.requiredMessage && <p innerHTML={this.requiredMessage}></p>}
         <slot></slot>
         {!this.readonly && !this.disabled && <slot name="actions"></slot>}