+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Release (2023-08-03-19-06_Unusual-Unmellow-Yellow-2)
+
+### Packages released
+
+- [img changelog](packages/img/CHANGELOG.md)
+- [mg-components changelog](packages/mg-components/CHANGELOG.md)
+
 ## Release (2023-07-25-16-29_Thankful-Turquoise-Blue-Tiger)
 
 ### Packages released